♻️  use UploadImage comp. in ChatScreen
